feat: Automatically generate chat title on first message
This commit enables automatic title generation when user send the first message on a chat.

We removed the `title` field on `POST /messages` route and a default `Novo chat` value is assigned to title until it gets generated.

The title is generated AFTER the assistant response to the user message as we need it to have more conversation context.